+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * cleaned up `app.register` method logic
 * using `koop.log` in index instead of `console.log`
 * reorganized index for readability & code organization
+* throw error on bad provider or cache registration
 
 ### Added
 * `app.registerProvider` method for providers